requirements.txt
akad==0.11.1
beautifulsoup4==4.9.3
certifi==2019.11.28
chardet==3.0.4
click==8.0.0
colorama==0.4.4
Flask==2.0.0
future==0.18.2
idna==2.8
itsdangerous==2.0.0
Jinja2==3.0.0
line-bot-sdk==1.19.0
linepy==3.0.8
MarkupSafe==2.0.0
pyasn1==0.4.8
PyQRCode==1.2.1
python-dotenv==0.10.3
requests==2.22.0
rsa==4.0
selenium==3.141.0
six==1.13.0
soupsieve==2.2.1
thrift==0.13.0
urllib3==1.25.7
Werkzeug==2.0.0
gunicorn==19.9.0